testui/requirements.txt
alabaster==0.7.12
astroid==2.4.2
attrs==19.3.0
autodoc==0.5.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
beautifultable==1.0.0
cached-property==1.5.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
clickhouse-driver==0.1.4
clickhouse-sqlalchemy==0.1.4
colorama==0.4.3
configparser==5.0.0
crayons==0.3.1
cryptography==3.0
decorator==4.4.2
deepdiff==5.0.1
delayed-assert==0.3.2
distro==1.5.0
dnspython==2.0.0
docker==4.2.2
docker-compose==1.26.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.16
et-xmlfile==1.0.1
Faker==4.1.1
google-api-core==1.22.2
google-auth==1.21.2
google-cloud-core==1.4.1
google-cloud-logging==1.15.1
googleapis-common-protos==1.52.0
grpcio==1.35.0
idna==2.10
imagesize==1.2.0
importlib-metadata==1.7.0
infi.clickhouse-orm==2.1.0
iso8601==0.1.12
isort==5.1.4
jdcal==1.4.1
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
lazy-object-proxy==1.4.3
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
meteo-downloader==1.0
more-itertools==8.4.0
mysql-connector-python==8.0.21
networkx==2.4
nose==1.3.7
numpy==1.19.5
openpyxl==3.0.4
ordered-set==4.0.2
packaging==20.4
pandas==1.0.5
paramiko==2.7.1
pdoc3==0.8.4
Pillow==8.0.1
pluggy==0.13.1
port-for==0.4
protobuf==3.12.2
psycopg2-binary==2.8.6
py==1.9.0
py-cpuinfo==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
pyjsparser==2.7.1
pylint==2.6.0
pymongo==3.10.1
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.3
pytest-assume==2.2.1
pytest-base-url==1.4.2
pytest-docker==0.7.2
pytest-docker-compose==3.1.2
pytest-docs==0.1.0
pytest-dotenv==0.5.2
pytest-html==2.1.1
pytest-metadata==1.10.0
pytest-selenium==1.17.0
pytest-variables==1.9.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-logstash==0.4.6
python-redmine==2.3.0
python3-logstash==0.4.80
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
selenium==3.141.0
six==1.15.0
snowballstemmer==2.0.0
soupsieve==2.0.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
splinter==0.13.0
SQLAlchemy==1.3.18
sqlalchemy-clickhouse==0.1.5.post0
tabulate==0.8.7
temlogger==0.4.0
text-unidecode==1.3
texttable==1.6.2
toml==0.10.1
toolz==0.10.0
tzlocal==2.1
urllib3==1.25.10
voluptuous==0.11.7
waitress==1.4.4
wcwidth==0.2.5
WebOb==1.8.6
websocket-client==0.57.0
WebTest==2.0.35
wrapt==1.12.1
zipp==3.1.0
